refactored swift_template.py
